+++ CHANGES.md
@@ -14,6 +14,7 @@ Release Notes.
 - Add the sub-command `endpoint get` to query single endpoint info.(#134)
 - Change the GraphQL method to the v9 version according to the server version.(#134)
 - Add `normal` field to Service entity.(#136)
+- Add the command `process` for query Process metadata.(#137)
 
 ### Bug Fixes